+++ tomb-common/src/utils/mod.rs
@@ -1,3 +1,3 @@
 pub mod error;
 pub mod serialize;
-pub mod tests;
+pub mod test;
